requirements.txt
fastapi==0.104.1
openai==1.3.5
pinecone-client==2.2.4
pgvector==0.2.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlmodel==0.0.14
uvicorn==0.24.0